# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "B2B API Doc",
        "description": "Api description...",
        "termsOfService": "",
        "contact": {
            "email": "la-ankan@mobiloitte.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/PROJECTS/b2b-and-ecommerce-platform-19103627-lamp/public",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/category": {
            "post": {
                "summary": "Category management",
                "description": "Category management",
                "operationId": "category",
                "consumes": [
                    "text/plain",
                    "application/json"
                ],
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Category management",
                        "required": true,
                        "schema": {
                            "properties": {
                                "user_id": {
                                    "description": "Category management",
                                    "type": "string",
                                    "default": "78"
                                },
                                "search": {
                                    "description": "Category management",
                                    "type": "string",
                                    "default": "null"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Category management",
                                    "type": "string",
                                    "default": "Data fetched succesfully"
                                },
                                "status": {
                                    "description": "Category management",
                                    "type": "string",
                                    "default": "1"
                                },
                                "result": {
                                    "description": "Category management",
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "description": "Category management",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "default": "1"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                }
            }
        },
...
},
...
}